Less Manual Obj Loading
A small change in code allows the name of the obj file to be inputed onto the python shell. Keep in mind you still have to type out, example: teapot.obj by full in order for the code to understand. In the next hours i'll be implementing a gui system (and trying to make obj viewing better lol)